chore: move builder, composer, css and pathmaker from pkg to internal
