+++ internal/builder/content_builder.go
