+++ internal/builder/ifile_content_builder.go
