+++ internal/builder/metadata_builder.go
