+++ internal/composer/folder_test.go
